vfs: use defaults for unknown metadata
If we augment an item but don't know the metadata, just
keep the values in the instantiated blank metadata, don't
override the uid/gid (are already None) and don't set the
time to the start, but keep it at the default of None.

This requires changing summary_bytes() a bit to show the
unknown value also in this case, not just when meta is
completely unknown.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>