test.sh: set LC_ALL=C for git fsck to avoid localized output
Observed on Ubuntu 20.04 with German as the primary language. The test
clearly parses output in English and since adapting the parsing to the
language of the test environment is not what the test tests,
specifying the language is the way to go.

The original problem can be reproduced via

  LC_ALL=de_DE.utf8  ./wvtest run t/test.sh

if your system has de_DE.utf8 avalilable.

Signed-off-by: Karl-Philipp Richter <krichter@posteo.de>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: adjust commit message; set LC_ALL=C just for
 fsck rather than setting LANGUAGE globally]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>