Use bup.repo.id for client index-cache/NAME when possible
If you have multiple repos on the same machine, or your host name
changes, or such a thing then the index-cache always gets very
confused and can end up downloading everything over and over again.

So whenever we can fetch a bup.repo.id from the remote, use that,
prefixed with "id--" so it can't collide with the existing names
(because they currently can't have a dash) for the client index-cache
directory name.

Also try to move the cache from the old location to the new one if the
old one exists, so the upgrade path doesn't re-download it all.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: rebase; handle servers without config-get]
[rlb@defaultvalue.org: add id-- prefix]
[rlb@defaultvalue.org: add release note]
[rlb@defaultvalue.org: adjust commit message]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>