cirrus: allow failures on macos with python 2
It looks like macos has changed something that makes it very difficult
to detect and use the *utime* related functions.  Apparently Emacs
just decided to live without it, and Python had to go to substantial
lengths to sort it out.  So until/unless we have a reasonable fix,
just mark that combination as expected to fail.

Python 3 should work fine since we rely on it to sort out the problem
for us.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>