Add path.indexcache to provide index cache location
Instead of finding the cache via git.repo('index-cache/...'), add a
path.indexcache() that takes just the '...' part as the argument, and
doesn't depend on an initialized git repo.

It currently defaults to the same location via a new
path.defaultrepo(), avoiding a circular dependency between bup.git and
bup.path. This will also make it possible to change the location (if
we consider this properly for automatic exclusion from indexing.)

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>