git: remove all_packdirs()
This really never made sense - if you have a local repository
and want to run midx/bloom on it, then it doesn't make sense
to touch the index caches for other remote repositories.

And if you want to operate manually on the cache for some odd
reason (it's maintained automatically) then you anyway need to
(and should) pass the --dir option for it explicitly.

Remove this to make things easier to reason about.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: previous behavior also didn't respect the
 documentation in bup-bloom(1).]