Teach pytest to handle WVSKIP and use it where we can
Have pytest scan the wvtest output for skip messages, and for now,
treat the whole test as skipped if we find any.  This means that
for now WVSKIP can only be used as an up-front guard, but that's all
we need in most cases.

This should help make it more likely that we'll notice when tests are
unexpectedly skipped.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>