Respect/prefer/remember user specified CPPFLAGS/CFLAGS/LDFLAGS
While we remembered a configure-time CC, we didn't do the same for
other common user specified build variables, i.e. CPPFLAGS, CFLAGS,
and LDFLAGS.  Fix that, and arrange for the user's values to come
last so that they can override the configuration values.

This matches expectations established by other tools, in particular
autoconf/automake, and avoids having to remember to specify overrides
to every invocation of make.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>