Avoid the use of a local window when splitting
The splitting code in `bupsplit.c` was written as if it had to handle
the case where bytes were fed one by-one, but it actually gets the whole
data chunk: there's no need to copy to a local window, when we have the
original data available: we can simply use the passed data to fetch the
byte that is leaving the current window.

Additionally, try to use __builtin_ctz for filling in the bits ptr, if
it is available.

Signed-off-by: Zachary Dremann <dremann@gmail.com>