configure.inc: fix LDFLAGS validation
Change the CFLAGS validation to create ngc$$ instead of ngc$$.o since
it doesn't specify -c.

Change the LDFLAGS validation to create the .o file it needs.
Previously it relied on the CFLAGS test, which was optional, to do
that, and the CFLAGS test was actually creating an executable, not an
object file.

Thanks to Greg Troxel for reporting the omission.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>