test-sparse-files: increase the test size for macos/afs
When the cirrus tests moved to monterey test-sparse-files started
failing in "sparse file restore --sparse (sparse middle)" because the
restore_size was 28 and so greater than the expected 20.

Guessing that this might just be due to higher fixed-costs, I doubled
the sparse region data_size, and that did allow the tests to pass.
While the diagnosis might or might not be correct, the change
shouldn't materially affect the tests' effectiveness.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>