Add file/folder illustration of the repo format
The DESIGN document did help me a lot to understand the functionality of bup quite a lot, however from reading "git for computer scientists" alone, I did not realize that git trees and blobs are 1-to-1 mapped to directories and files on `git checkout`. I searched for ways in how git alone can merge several blobs to one file, which it cant. The reason is this sentence:

> ... so you can use git to manipulate the bup repository if you want, and you
> probably won't break anything.  It's also a comfort to know you can squeeze
> data out using git, just in case bup fails you, and as a developer, git offers
> some nice tools ...

where I interpreted "squeeze out" as a simple "git checkout", which is not true, as bup
splits files. The added text makes this more concrete and prepares the reader for the
parts about hashsplitting and fanout

Signed-off-by: Moritz Lell <mlell08@gmail.com>