Suppress the git main branch name message on bup init
If the git config option init.defaultBranch is not set, git init
prints a message on how to change the default branch name. The default
branch name is never used, so the message is irrelevant and might be
confusing for users.  Setting some explicit default branch name
suppresses that warning.

The message might not appear for most frequent git users, as they will
have set git config --global init.defaultBranch at some point to get
rid of this message on git init. However, it will affect most new
users and users that use only bup, not git. It is therefore an UX
improvement.

Signed-off-by: Moritz Lell <mlell08@gmail.com>