Fix ACL metadata format; delimit short form entries with commas
Previously, ACL entries were delimited by newlines, which is incorrect
according to acl(5), and the POSIX1e standard, and which is rejected
with EINVAL by acl_from_text() on some platforms (e.g. cygwin).

To fix the problem, change read_acl to use comma delimiters when
reading new ACLs for a path, and write those in a new "v2" ACL
metadata record.  When loading the previous newline delimited v1
records, convert newline delimiters to commas unless the ACL already
contains a comma, in which case, defer an error and ignore the entire
ACL record.

Of course, if any recorded ACL field contained a newline (in the user
or group name, or the content of any custom fields), then this
conversion might be trouble.  (Debian, at least, forbids newlines in
user or group names.)

And even with the standard-specified commas, the standard also doesn't
appear to say anything about what should happen if a field value
contains a comma.  So in the longer run, it looks like we should stop
relying on the built-in text formats.

Thanks to Moritz Lell for reporting the problem.

cf. (via acl(5)) http://wt.tuxomania.net/publications/posix.1e/download.html

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>