correct_posix1e_v1_delimiters: provide path for error messages
The previous invocation omitted the path argument.  Include it, and
adjust the code to accommodate the fact that the Metadata path is
optional.

Adjust test-meta-acls to also validate restoration of saves created by
BUP_TEST_OTHER_BUP when it's set.  That catches the problem when the
other bup is, for example, 0.33.

cf. 5772da806f6b008ee7fbbef19b56c2bb7ed85703

Thanks to Phil Sutter and Johannes Berg (at least) for reporting the
problem.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>