fuse: don't filter output when backgrounding
This prevented a backgrounded (non --foreground) "bup fuse" from
exiting after an unmount.  Instead it would block forever on the
attempt to join the filter_thread which no longer existed after
libfuse forked while daemonizing the process.

For now, just add fuse to the list of commands that skip output
filtering.  If nothing else, we may eventually want it to be
predicated on the --foreground status.

Thanks to Rob Leslie for reporting the problem:
https://bugs.debian.org/cgi-bin/1050040

Tested-by: Rob Browning <rlb@defaultvalue.org>
Signed-off-by: Rob Browning <rlb@defaultvalue.org>