update-checkout-info: simplify check for git tree (don't use git)
The previous rev-parse check would print warnings when there was no
git tree, and since we've already attempted to make sure we're at the
top of a bup tree via the bupsplit.c test, just looking for a ./.git
dir or file (worktrees) should be sufficient.

Thanks to Alain Cochard for reporting the issue.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>