bup: helpers: significantly speed up columnate()
I was accidentally running 'bup ls' on a directory in my backup
that has close to 600k files in it, and it just didn't finish,
even after close to an hour.

Turns 'bup ls -l' is just fine, and columnate() is the issue,
and debugging that a bit further all the string building really
is the issue.

Change this to build strings in the recommended way by joining
a list of strings, rather than keep appending an existing one,
which is really easy in this case.

Now 'bup ls' on the directory finishes in about 35 seconds.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>
(cherry picked from commit 432bd2678294ffb87b33783010d166b41b5ba65f)