pytest: default to --dist worksteal when possible
Default to worksteal if the version's new enough since it claims, and
appears, to handle test sets with more widely varying run times
better.

Move the xdist detection and other pytest default argument selection
to python.  This both puts it all in one place, and avoids extra
python invocations testing for suitability.  As part of that, just
have ./pytest drop xdist options (i.e. -n) when it's not available.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>