fsck: run par2 generate in a sandbox to avoid empty files
If par2 generate is interrupted, it may leave empty *.par2 files, and
not just the top-level PATH.par2 index.  To avoid that possiblity, run
par2 in a temporary directory and then move the recovery files to the
right place if par2 succeeds.

fsck currently checks whether the top-level index is empty, but
doesn't check any other vol files.

Thanks to Johannes Berg for reporting the problem.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>