configure.inc: fix main() prototype and other (now) invalid test code
Some newer compilers won't accept a main without a return type or
argument types, and there's no reason to be doing that, so don't.

Improve the function and field checks to avoid newer warnings and
errors, e.g. just test linkage, so we don't have to worry (yet) about
which headers a given platform needs.

Thanks to Johannes Berg for help devising the improvements.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>